#156 Rename utility module for simulator
