+++ neclib/utils/__init__.py
@@ -3,6 +3,6 @@
 """Collection of general purpose functionalities."""
 
 from .data_utils import *
-from .function_utils import *
+from .simulator_utils import *
 from .math_utils import *
 from .quantity_utils import *
